app/requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==6.0.0
gunicorn==21.2.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.8.0
pyscopg2==66.0.2
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
